app/requirements.txt
google-adk==1.21.0
litellm==1.80.10
python-dotenv
pydantic
pydantic-settings
greenlet

# Investment agent dependencies
kiteconnect>=4.3.0
yfinance>=0.2.40
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
pyotp>=2.9.0

# Trade agent dependencies
numpy>=1.24.0
TA-Lib>=0.4.28